improve local store serialization, closes #406
* Turn "msgpack" decoding to asynchronous mode.
* Serialize/deserialize mails by portions, one by one, so "msgpack" never faces too huge bunch of bytes. This also means that memory use peaks get reduced (one by one portions processing).
* Compress the stored data (gzip).
* Compress the mails "raw" and "body" props (lzutf8).
* Double the default app starting memory related arguments.

+++ src/electron-main/api/index.ts
@@ -129,8 +129,8 @@ export const initApi = async (ctx: Context): Promise<IpcMainApiEndpoints> => {
         },
 
         // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-        async logout() {
-            if (ctx.keytarSupport) {
+        async logout({skipKeytarProcessing}) {
+            if (!skipKeytarProcessing && ctx.keytarSupport) {
                 await deletePassword();
             }
